add arena module and remove misc
