+++ data/spawner.ron
@@ -4,15 +4,15 @@
 			formation_spawnables: [
 				(
 					spawnable_type: Mob(Enemy(MissileLauncher)),
-					position: (0.0, 70.0),
+					position: (0.0, 50.0),
 				),
 				(
 					spawnable_type: Mob(Enemy(Pawn)),
-					position: (20.0, 80.0),
+					position: (20.0, 60.0),
 				),
 				(
 					spawnable_type: Mob(Enemy(Pawn)),
-					position: (-20.0, 80.0),
+					position: (-20.0, 60.0),
 				),
 			],
 			weight: 1.0,
@@ -22,15 +22,15 @@
 			formation_spawnables: [
 				(
 					spawnable_type: Mob(Enemy(Drone)),
-					position: (0.0, 70.0),
+					position: (0.0, 50.0),
 				),
 				(
 					spawnable_type: Mob(Ally(Hauler)),
-					position: (20.0, 80.0),
+					position: (20.0, 60.0),
 				),
 				(
 					spawnable_type: Mob(Ally(Hauler)),
-					position: (-20.0, 80.0),
+					position: (-20.0, 60.0),
 				),
 			],
 			weight: 1.0,
