+++ src/arena/mod.rs
@@ -1,0 +1,7 @@
+mod barrier;
+mod gate;
+
+pub use self::{
+    barrier::spawn_barriers_system,
+    gate::{despawn_gates_system, spawn_despawn_gates_system},
+};
