--- src/misc/mod.rs
@@ -1,5 +1,0 @@
-//! `thetawave_lib` misc module
-
-mod spawn;
-
-pub use self::spawn::{despawn_spawnables_system, spawn_barrier_system};
